+++ index.html
@@ -5,6 +5,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <script type="module" src="js/main.js">
   <title>Кексобукинг</title>
 </head>
 <body>
